Remove utils dir and move commons to helpers
