+++ src/core/controllers/mouse-gesture-controller.mjs
@@ -1,4 +1,4 @@
-import { toSingleButton, getDistance } from "/core/utils/commons.mjs";
+import { toSingleButton, getDistance } from "/core/helpers/commons.mjs";
 
 // global static variables
