+++ src/core/controllers/wheel-gesture-controller.mjs
@@ -1,4 +1,4 @@
-import { toSingleButton } from "/core/utils/commons.mjs";
+import { toSingleButton } from "/core/helpers/commons.mjs";
 
 // global static variables
