+++ src/core/models/pattern-constructor.mjs
@@ -1,4 +1,4 @@
-import { vectorDirectionDifference } from "/core/utils/commons.mjs";
+import { vectorDirectionDifference } from "/core/helpers/commons.mjs";
 
 
 /**
