+++ src/core/views/mouse-gesture-view/mouse-gesture-view.mjs
@@ -1,4 +1,4 @@
-import { getDistance } from "/core/utils/commons.mjs";
+import { getDistance } from "/core/helpers/commons.mjs";
 
 /**
  * MouseGestureView "singleton"
