+++ src/core/views/popup-command-view/popup-command-view.mjs
@@ -1,4 +1,4 @@
-import { isEmbeddedFrame } from "/core/utils/commons.mjs";
+import { isEmbeddedFrame } from "/core/helpers/commons.mjs";
 
 
 /**
